Move awaitGC to common utility package
