+++ dd-java-agent/agent-tooling/agent-tooling.gradle
@@ -19,6 +19,7 @@ dependencies {
 
   testCompile deps.opentracing
   testCompile project(':dd-java-agent:testing')
+  testCompile project(':utils:gc-utils')
 
   instrumentationMuzzle sourceSets.main.output
   instrumentationMuzzle configurations.compile
