+++ dd-java-agent/dd-java-agent.gradle
@@ -92,6 +92,7 @@ modifyPom {
 dependencies {
   testCompile project(':dd-trace-api')
   testCompile project(':dd-trace-ot')
+  testCompile project(':utils:gc-utils')
 
   testCompile deps.opentracingMock
   testCompile deps.testLogging
