+++ dd-java-agent/testing/src/test/java/muzzle/MuzzleWeakReferenceTest.java
@@ -1,9 +1,9 @@
 package muzzle;
 
-import datadog.trace.agent.test.TestUtils;
 import datadog.trace.agent.tooling.muzzle.Reference;
 import datadog.trace.agent.tooling.muzzle.ReferenceCreator;
 import datadog.trace.agent.tooling.muzzle.ReferenceMatcher;
+import datadog.trace.util.gc.GCUtils;
 import java.lang.ref.WeakReference;
 import java.net.URL;
 import java.net.URLClassLoader;
@@ -26,7 +26,7 @@ public class MuzzleWeakReferenceTest {
     final ReferenceMatcher refMatcher = new ReferenceMatcher(refs);
     refMatcher.getMismatchedReferenceSources(loader);
     loader = null;
-    TestUtils.awaitGC(clRef);
+    GCUtils.awaitGC(clRef);
     return clRef.get() == null;
   }
 }
