+++ dd-trace-ot/dd-trace-ot.gradle
@@ -38,6 +38,7 @@ dependencies {
   testImplementation deps.autoservice
 
   testCompile project(":dd-java-agent:testing")
+  testCompile project(':utils:gc-utils')
   testCompile group: 'org.assertj', name: 'assertj-core', version: '2.9.+'
   testCompile group: 'org.mockito', name: 'mockito-core', version: '2.19.0'
   testCompile group: 'org.objenesis', name: 'objenesis', version: '2.6'
