+++ dd-trace/dd-trace.gradle
@@ -5,15 +5,14 @@ dependencies {
   annotationProcessor deps.autoservice
   implementation deps.autoservice
 
-  compile project(':dd-java-agent:agent-bootstrap')
   compile project(':dd-trace-api')
 
-  compile deps.jackson
   compile deps.slf4j
-  // any higher versions seems to break ES tests with this exception:
-  // java.lang.NoSuchMethodError: com.fasterxml.jackson.dataformat.smile.SmileGenerator.getOutputContext()
-  compile group: 'org.msgpack', name: 'jackson-dataformat-msgpack', version: '0.8.14'
+  compile deps.jackson
+  compile group: 'org.msgpack', name: 'jackson-dataformat-msgpack', version: '0.8.16'
 
+  compile project(':utils:gc-utils')
+  
   // Spock uses that for mocking
   testCompile deps.bytebuddy
   testCompile group: 'org.objenesis', name: 'objenesis', version: '2.6' // Last version to support Java7
