+++ settings.gradle
@@ -16,6 +16,7 @@ include ':dd-java-agent:agent-jmxfetch'
 
 // misc
 include ':dd-java-agent:testing'
+include ':utils:gc-utils'
 
 // smoke tests
 include ':dd-smoke-tests:play'
