+++ utils/gc-utils/gc-utils.gradle
@@ -1,0 +1,1 @@
+apply from: "${rootDir}/gradle/java.gradle"
